Create a file for lint_module_test
Also recover three functions that were clearly class function for
LintModuleTest.
